Lesson 1 -2 assignment
