+++ lesson1/src/AddTodoForm.js
@@ -1,0 +1,21 @@
+import React from "react"
+
+
+
+function AddTodoForm(){
+    return(
+        <form>
+            <label htmlFor="todoTitle">
+                Title
+            </label>
+            <input id="todoTitle">
+
+            </input>
+            <button>
+                Add
+            </button>
+        </form>
+    )
+}
+
+export default AddTodoForm;
